kingst-la2016: style nits in FPGA code path and threshold selection
Check sigrok routine return codes for SR_OK not zero. The snprintf(3)
routine does include the NUL termination in the length spec. Use size_t
for the threshold selection which in essence is an enumeration.

+++ src/hardware/kingst-la2016/protocol.h
@@ -108,7 +108,7 @@ struct dev_context {
 
 	/* User specified parameters. */
 	struct pwm_setting pwm_setting[LA2016_NUM_PWMCH_MAX];
-	unsigned int threshold_voltage_idx;
+	size_t threshold_voltage_idx;
 	float threshold_voltage;
 	uint64_t max_samplerate;
 	uint64_t cur_samplerate;
